requirements.txt
kafka-python
requests
csv
json
mysql-connector-python
```